streaming API attempt

+++ mp3-extract/src/main/java/su/drei/mp3extr/exporter/IDataExporter.java
@@ -4,7 +4,7 @@ public interface IDataExporter {
 
     void init(float sampleRate, int channels);
 
-    void exportPcmBatch(int channel, int[] pcm);
+    void exportPcmBatch(int channel, double[] pcm);
 
     void exportFrequencyDomainBatch(int channel, double[] freqDomainBatch);
